requirements.txt
# Core Framework
streamlit==1.31.0

# LangChain & RAG
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5

# Vector Store
faiss-cpu==1.7.4
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1
openai==1.10.0

# PDF Processing
PyMuPDF==1.23.21
unstructured==0.12.2
pdfplumber==0.10.3

# Text Processing
tiktoken==0.5.2

# HTTP & API
requests==2.31.0
httpx==0.26.0

# Data Handling
pandas==2.2.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
tqdm==4.66.1

# ML Features (Advanced Functionalities)
transformers==4.36.2
scikit-learn==1.4.0
torch==2.1.2
